Refactor: Header into file
